+++ freesms/__init__.py
@@ -10,7 +10,7 @@ See:
 
 import requests
 
-__version__ = '0.2.0'
+__version__ = '0.2.1'
 
 
 class FreeResponse(object):